refactor:페이지별로 맡는 역할 재분배
